Refactor API boostrap into separate functions
