+++ api/src/common/helper/setup-logger.ts
@@ -1,0 +1,10 @@
+import { NestExpressApplication } from '@nestjs/platform-express';
+import { AppLogger } from 'src/modules/logger/app-logger';
+
+export const setupLogger = (app: NestExpressApplication): AppLogger => {
+  const logger = app.get(AppLogger);
+  app.useLogger(logger); // Nest framework logs go through our custom logger
+  app.flushLogs();
+
+  return logger;
+};
